update SDK version to 1.1.0
- add new domain apis
- add new versions for commodity apis
- fix schema for some objects

+++ src/apifreaks/types/email_validate_response_address_security.py
@@ -7,7 +7,7 @@ from ..core.pydantic_utilities import IS_PYDANTIC_V2, UniversalBaseModel
 
 
 class EmailValidateResponseAddressSecurity(UniversalBaseModel):
-    threat_score: int
+    threat_score: float
     is_tor: bool
     is_proxy: bool
     proxy_type: str
